restricted edit/delete functions to users who post them

+++ chirper/views.py
@@ -1,6 +1,9 @@
 from django.shortcuts import render, redirect
 from .models import Post
 from .forms import PostForm
+from django.contrib.auth.decorators import login_required
+from django.http import Http404
+
 # Create your views here.
 
 def index(request):
@@ -8,12 +11,16 @@ def index(request):
     context = {"posts": posts}
     return render(request, 'chirper/index.html',  context)
 
+
+@login_required
 def new_post(request):
     if request.method != 'POST':
         form = PostForm()
     else:
         form = PostForm(data=request.POST)
         if form.is_valid:
+            new_post = form.save(commit=False)
+            new_post.poster = request.user
             form.save()
             return redirect('chirper:index')
     
@@ -21,24 +28,32 @@ def new_post(request):
     return render(request, 'chirper/new_post.html', context)
 
 
+@login_required
 def edit_post(request, post_id):
     post = Post.objects.get(id=post_id)
-    
-    if request.method != 'POST':
-        form = PostForm(instance=post)
+    if post.poster != request.user:
+        raise Http404
     else:
-        form = PostForm(instance=post, data=request.POST)
-        if form.is_valid:
-            form.save()
-            return redirect('chirper:index')
+        if request.method != 'POST':
+            form = PostForm(instance=post)
+        else:
+            form = PostForm(instance=post, data=request.POST)
+            if form.is_valid:
+                form.save()
+                return redirect('chirper:index')
     context = {'post': post, 'form':form}
     return render(request, 'chirper/edit_post.html', context)
 
+
+@login_required
 def delete_post(request, post_id):
     post = Post.objects.get(id=post_id)
     context = {'post' : post}
-    if request.method != 'POST':
-        return render(request, 'chirper/edit_post.html', context)
-    else: 
-        post.delete()   
-        return redirect('chirper:index')
+    if post.poster == request.user:
+        if request.method != 'POST':
+            return render(request, 'chirper/edit_post.html', context)
+        else: 
+            post.delete()   
+            return redirect('chirper:index')
+    else:
+        raise Http404
